+++ src/MOCKernel.h
@@ -66,7 +66,7 @@ public:
   virtual void newTrack(Track* track);
 
   /* Executing function describes kernel behavior */
-  virtual void execute(double length, Material* mat, int fsr_id,
+  virtual void execute(double length, Material* mat, long fsr_id,
                        int track_idx, int cmfd_surface_fwd,
                        int cmfd_surface_bwd, double x_start, double y_start,
                        double z_start, double phi, double theta)=0;
@@ -86,7 +86,7 @@ class CounterKernel: public MOCKernel {
 
 public:
   CounterKernel(TrackGenerator* track_generator, int row_num);
-  void execute(double length, Material* mat, int fsr_id,
+  void execute(double length, Material* mat, long fsr_id,
                int track_idx, int cmfd_surface_fwd, int cmfd_surface_bwd,
                double x_start, double y_start, double z_start,
                double phi, double theta);
@@ -122,7 +122,7 @@ public:
 
   VolumeKernel(TrackGenerator* track_generator, int row_num);
   void newTrack(Track* track);
-  void execute(double length, Material* mat, int fsr_id,
+  void execute(double length, Material* mat, long fsr_id,
                int track_idx, int cmfd_surface_fwd, int cmfd_surface_bwd,
                double x_start, double y_start, double z_start,
                double phi, double theta);
@@ -146,7 +146,7 @@ private:
 
 public:
   SegmentationKernel(TrackGenerator* track_generator, int row_num);
-  void execute(double length, Material* mat, int fsr_id,
+  void execute(double length, Material* mat, long fsr_id,
                int track_idx, int cmfd_surface_fwd, int cmfd_surface_bwd,
                double x_start, double y_start, double z_start,
                double phi, double theta);
@@ -192,7 +192,7 @@ public:
                     int track_id);
   void setTrackIndexes(int azim_index, int polar_index);
   void setDirection(bool direction);
-  void execute(double length, Material* mat, int fsr_id,
+  void execute(double length, Material* mat, long fsr_id,
                int track_idx, int cmfd_surface_fwd, int cmfd_surface_bwd,
                double x_start, double y_start, double z_start,
                double phi, double theta);
